kinwalletservice/requirements.txt
setuptools==37.0.0
datadog
psycopg2==2.7.3.2
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.1.14
lxml==4.1.1
Flask-Cors==3.0.3
grequests==0.3.0
requests==2.3.0
simplejson==3.12.0
python-redis-lock==3.2.0
redis==2.10.6
boto3==1.4.4
